refactor(saves): decompose SavesTab.tsx into saves/ sub-folder
Splits the 1104-LOC SavesTab.tsx along the author-marked section
banners into a saves/ sub-folder. No behavior change.

  - helpers.ts (pure formatters + selectors)
  - NewSlotModal.tsx
  - VersionHistoryPanel.tsx
  - SaveFileRow.tsx
  - ServerSaveRow.tsx
  - SlotPanel.tsx

SavesTab.tsx drops from 1104 to ~250 LOC. Pure helpers gain tests.

Closes #613
